Move AWS SDK related classes to a SDK package
