+++ src/test/java/cloud/fogbow/ras/core/plugins/interoperability/aws/compute/v2/AwsComputePluginTest.java
@@ -12,6 +12,7 @@ import java.util.TreeSet;
 
 import cloud.fogbow.common.exceptions.*;
 import cloud.fogbow.common.util.BinaryUnit;
+import cloud.fogbow.ras.core.plugins.interoperability.aws.sdk.compute.model.AwsHardwareRequirements;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
